chore: consolidate and remove unnecessary icons
