--- src/icons/SmoothX.tsx
@@ -1,36 +1,0 @@
-import * as React from 'react'
-import Colors from 'src/styles/colors'
-import Svg, { Path } from 'svgs'
-
-interface Props {
-  height?: number
-  color?: string
-}
-
-export default class SmoothX extends React.PureComponent<Props> {
-  static defaultProps = {
-    height: 47,
-    color: Colors.white,
-  }
-
-  render() {
-    return (
-      <Svg
-        xmlns="http://www.w3.org/2000/svg"
-        height={this.props.height}
-        width={this.props.height}
-        viewBox={`0 0 92.132 92.132`}
-        fill="none"
-      >
-        <Path
-          d="M2.141,89.13c1.425,1.429,3.299,2.142,5.167,2.142c1.869,0,3.742-0.713,5.167-2.142l33.591-33.592L79.657,89.13
-          c1.426,1.429,3.299,2.142,5.167,2.142c1.867,0,3.74-0.713,5.167-2.142c2.854-2.854,2.854-7.48,0-10.334L56.398,45.205
-          l31.869-31.869c2.855-2.853,2.855-7.481,0-10.334c-2.853-2.855-7.479-2.855-10.334,0L46.065,34.87L14.198,3.001
-          c-2.854-2.855-7.481-2.855-10.333,0c-2.855,2.853-2.855,7.481,0,10.334l31.868,31.869L2.143,78.795
-          C-0.714,81.648-0.714,86.274,2.141,89.13z"
-          fill={this.props.color}
-        />
-      </Svg>
-    )
-  }
-}
